♻️ [refactor]: #452 Phase 1: lexer_new_position.rs を分離
new / position の 3 テストを `src/lexer/tests/lexer_new_position.rs` に
分離。`tests.rs` に `mod lexer_new_position;` 宣言を追加し本体から
該当 Phase の `#[test]` 群と Phase 区切りコメントを削除。

- 293 件 グリーン維持
- 短縮テスト関数名集合 diff 0 行 / 件数 293 / 重複 0
- cargo fmt / clippy グリーン
- lexer_new_position.rs: 24 行 (< 800)

Refs #452
